Organzie random_graph
