+++ rustworkx/matching.pyi
@@ -7,11 +7,11 @@
 # that they have been altered from the originals.
 
 # This file contains only type annotations for PyO3 functions and classes
-# For implementation details, see __init__.py and src/lib.rs
+# For implementation details, see __init__.py and src/matcing/mod.rs
 
 from .iterators import *
-from .graph import PyGraph as PyGraph
-from .digraph import PyDiGraph as PyDiGraph
+from .graph import PyGraph
+from .digraph import PyDiGraph
 
 from typing import Optional, Set, TypeVar, Tuple, Callable
